Make the platform library targets OBJECT libraries
Defining the library target as STATIC will cause linking issues
where the WEAK implementation of a function is taken when both
weakly and strongly defined implementations of the same function exist.

This is due to the difference in how the linker treats .o files vs .a files.

Making the platform library an OBJECT library target is essentially
the equivalent of using the -Wl,--whole-archive flag when linking against
a STATIC library (.a)