[github-actions] move DockerHub publish job into a separate workflow
Also, this fixes the bug where the DockerHub username and password
secrets are unavailable in the DockerHub login step.

[This
failure](https://github.com/SiliconLabs/ot-efr32/actions/runs/9681942492/job/26713977695)
is caused by a bug I accidentally introduced in
https://github.com/openthread/ot-efr32/pull/896. The secrets consumed by
a reusable workflow must be defined as inputs.

See this for more details
https://docs.github.com/en/actions/sharing-automations/reusing-workflows#using-inputs-and-secrets-in-a-reusable-workflow